Implement Votable and Created interfaces
