+++ src/main/kotlin/net/model/RedditStructure.kt
@@ -5,6 +5,7 @@ import kotlinx.serialization.Serializable
 import kotlinx.serialization.builtins.ListSerializer
 import kotlinx.serialization.json.*
 
+
 /** Describes the data structure of objects returned when using the Reddit JSON API. */
 
 interface Votable {
@@ -37,3 +38,68 @@ interface Created {
     val createdUtc: Long
 }
 
+@Serializable
+sealed class Thing {
+    /** This item's identifier, e.g. "8xwlg" */
+    abstract val id: String
+
+    /** Fullname of comment, e.g. "t1_c3v7f8u" */
+    abstract val name: String
+}
+
+@Serializable
+data class Comment(
+    override val id: String,
+    override val name: String,
+    override val ups: Int,
+    override val downs: Int,
+    override val likes: Boolean?,
+    override val created: Long,
+    override val createdUtc: Long,
+
+
+    ) : Thing(), Votable, Created
+
+@Serializable
+data class Link(
+    override val id: String,
+    override val ups: Int,
+    override val downs: Int,
+    override val likes: Boolean?,
+    override val created: Long,
+    override val createdUtc: Long,
+    override val name: String,
+
+
+    ) : Thing(), Votable, Created
+
+
+/**
+ * A listing response contains many [Thing], which can be [Comment] or [Link]. This object takes a [JsonArray]
+ * of [Thing] and returns a [List] of [Thing], properly classified into [Comment] or [Link].
+ */
+object PolymorphicThingUnwrapper : JsonTransformingSerializer<List<Thing>>(ListSerializer(Thing.serializer())) {
+
+    /** @param element A [JsonArray] of [Thing] */
+    override fun transformDeserialize(element: JsonElement): JsonElement {
+        require(element is JsonArray)
+        val thingArray = element.jsonArray
+
+        return JsonArray(thingArray.map {
+            JsonObject(
+                it.jsonObject["data"]!!.jsonObject.toMutableMap().apply {
+                    // Add a "type" to the JSON object to preserve the class discriminator.
+                    // This allows seamless polymorphism.
+                    // Reddit classifies it's data structures according to type by using a prefix.
+                    // t1 means that the [Thing] is a [Comment], t2 means [Link]
+                    this["type"] =
+                        if (it.jsonObject["kind"]!!.jsonPrimitive.content == "t1") {
+                            JsonPrimitive("net.model.Comment")
+                        } else {
+                            JsonPrimitive("net.model.Link")
+                        }
+                }
+            )
+        })
+    }
+}
